refactor: main.js에서 로직 별로 분리
