+++ src/InitHeader.js
@@ -1,0 +1,6 @@
+import Header from "./components/Header";
+
+export default function InitHeader() {
+  const $headerContainer = document.querySelector(".gnb");
+  Header($headerContainer);
+}
