+++ src/InitRestaurantList.js
@@ -1,0 +1,8 @@
+import RestaurantList from "./components/RestaurantList";
+
+export default function InitRestaurantList() {
+  const $restaurantListContainer = document.querySelector(
+    ".restaurant-list-container"
+  );
+  RestaurantList($restaurantListContainer);
+}
